fix sessions, app bar, email

+++ apps/crud/views.py
@@ -32,14 +32,11 @@ class EmailView(TemplateView):
 
     def get(self, request):
         try:  
-            form = EmailForm()
-            excepcion.enviar_mensaje(
-                'Nuevo Email %s' % (form.cleaned_data('email')),
-                '%s' % (str(e))
-            )
-            
-            return render(request, self.template_name, {'form':form})
-
+            if 'email' in request.session:
+                return HttpResponseRedirect('/empleados/')
+            else:
+                form = EmailForm()  
+                return render(request, self.template_name, {'form':form})
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.EmailView %s method:get' % (self.template_name),
@@ -51,6 +48,7 @@ class EmailView(TemplateView):
             form = EmailForm(request.POST)
             if form.is_valid():
                 email = form.cleaned_data['email']
+                request.session['email'] = email
                 return HttpResponseRedirect('/empleados/')
         except Exception as e:
             excepcion.enviar_mensaje(
@@ -63,8 +61,12 @@ class HomeView(TemplateView):
 
     def get(self, request):
         try:
-            empleados = Empleado.objects.order_by('-fecha_registro')
-            return render(request,self.template_name,{'empleados':empleados})    
+            if 'email' in request.session:
+                email = request.session['email']
+                empleados = Empleado.objects.order_by('-fecha_registro')
+                return render(request,self.template_name,{'empleados':empleados, 'email': email})
+            else:
+                return HttpResponseRedirect("/")   
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.HomeView %s method:get' % (self.template_name),
@@ -76,8 +78,12 @@ class RegistrarEmpleadoView(TemplateView):
 
     def get(self, request):
         try:
-            form = EmpleadoForm()
-            return render(request, self.template_name,{'form':form} )
+            if 'email' in request.session:
+                email = request.session['email']
+                form = EmpleadoForm()
+                return render(request, self.template_name,{'form':form, 'email' : email} )
+            else:
+                return HttpResponseRedirect("/")   
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.RegistrarEmpleadoView %s method:get' % (self.template_name),
@@ -86,20 +92,23 @@ class RegistrarEmpleadoView(TemplateView):
        
     def post(self, request):
         try:
-            if request.method == 'POST':
-                form = EmpleadoForm(request.POST)
-                if form.is_valid():
-                    empleado = Empleado()
-                    empleado.nombre = form.cleaned_data['nombre']
-                    empleado.a_materno = form.cleaned_data['a_materno']
-                    empleado.a_paterno = form.cleaned_data['a_paterno']
-                    empleado.fecha_nacimiento = form.cleaned_data['fecha_nacimiento']
-                    empleado.fecha_registro = timezone.now()
-                    empleado.sexo = True
-                    empleado.save()
-                    return HttpResponseRedirect('/empleados')
-                else:
-                    return render(request, self.template_name, {'resultados' : 'El formulario no es valido'})
+            if 'email' in request.session:
+                if request.method == 'POST':
+                    form = EmpleadoForm(request.POST)
+                    if form.is_valid():
+                        empleado = Empleado()
+                        empleado.nombre = form.cleaned_data['nombre']
+                        empleado.a_materno = form.cleaned_data['a_materno']
+                        empleado.a_paterno = form.cleaned_data['a_paterno']
+                        empleado.fecha_nacimiento = form.cleaned_data['fecha_nacimiento']
+                        empleado.fecha_registro = timezone.now()
+                        empleado.sexo = True
+                        empleado.save()
+                        return HttpResponseRedirect('/empleados')
+                    else:
+                        return render(request, self.template_name, {'resultados' : 'El formulario no es valido, revise los datos que está tratando de ingresar'})
+            else:
+                return HttpResponseRedirect("/")   
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.RegistrarEmpleadoView %s method:post' % (self.template_name),
@@ -111,41 +120,74 @@ class EditarEmpleadoView(TemplateView):
 
     def get(self, request, pk):
         try:
-            empleado = Empleado.objects.get(pk = pk)
-            form = EmpleadoForm(initial={
-                'nombre' : empleado.nombre, 
-                'a_paterno' : empleado.a_paterno, 
-                'a_materno' :  empleado.a_materno, 
-                'fecha_nacimiento' : empleado.fecha_nacimiento,})
-            return render(request, self.template_name, {'form' : form,})
+            if 'email' in request.session:
+                email = request.session['email']
+                empleado = Empleado.objects.get(pk = pk)
+                form = EmpleadoForm(initial={
+                    'nombre' : str.strip(empleado.nombre), 
+                    'a_paterno' : str.strip(empleado.a_paterno), 
+                    'a_materno' :  str.strip(empleado.a_materno), 
+                    'fecha_nacimiento' : empleado.fecha_nacimiento,})
+                return render(request, self.template_name, 
+                    {
+                        'form' : form,
+                        'email' : email
+                    }
+                )
+            else:
+                return HttpResponseRedirect("/")
         except Empleado.DoesNotExist:    
             return render(request, 'crud/excepciones.html', {'error' : 'Emplead@ no existente'})
 
     def post(self, request, pk):
         try:
-            if request.method == 'POST':
-                form = EmpleadoForm(request.POST)
-                if form.is_valid():
-                    try:
-                        empleado = get_object_or_404(Empleado,pk = pk)
-                        empleado.nombre = form.cleaned_data['nombre']
-                        empleado.a_paterno = form.cleaned_data['a_paterno']
-                        empleado.a_materno = form.cleaned_data['a_materno']
-                        empleado.fecha_nacimiento = form.cleaned_data['fecha_nacimiento']
-                        empleado.save()
-                        
-                        form = EmpleadoForm(initial={
-                            'nombre' : empleado.nombre, 
-                            'a_paterno' : empleado.a_paterno, 
-                            'a_materno' :  empleado.a_materno, 
-                            'fecha_nacimiento' : empleado.fecha_nacimiento,}
-                        )
+            if 'email' in request.session:
+                email = request.session['email']
 
-                        return render(request, self.template_name, {'resultado' : 'Se ha actualizado el registro exitosamente','form' :form},)
-                    except Empleado.DoesNotExist:
-                        return render(request, 'crud/excepciones.html', {'error' : 'El usuario que se intenta actualizar no existe'})
-                else:
-                    return render(request, self.template_name, {'resultado' : 'Se ha actualizado el registro exitosamente'})
+                if request.method == 'POST':
+                    form = EmpleadoForm(request.POST)
+                    if form.is_valid():
+                        try:
+                    
+                            empleado = get_object_or_404(Empleado,pk = pk)
+                            empleado.nombre = str.strip(form.cleaned_data['nombre'])
+                            empleado.a_paterno = str.strip(form.cleaned_data['a_paterno'])
+                            empleado.a_materno = str.strip(form.cleaned_data['a_materno'])
+                            empleado.fecha_nacimiento = form.cleaned_data['fecha_nacimiento']
+                            empleado.save()
+                            
+                            form = EmpleadoForm(initial={
+                                'nombre' : empleado.nombre, 
+                                'a_paterno' : empleado.a_paterno, 
+                                'a_materno' :  empleado.a_materno, 
+                                'fecha_nacimiento' : empleado.fecha_nacimiento,}
+                            )
+
+                            return render(request, self.template_name, 
+                                {
+                                    'resultado' : 'Se ha actualizado el registro exitosamente',
+                                    'form' :form,
+                                    'email':email
+                                },
+                            )
+                        except Empleado.DoesNotExist:
+                            return render(request, 'crud/excepciones.html', 
+                                {
+                                    'error' : 'El usuario que se intenta actualizar no existe',
+                                    'email' : email
+                                }
+                            )
+                    else:
+                        return render(request, self.template_name, 
+                            {
+                                'resultado' : 'El formulario no es valido, revise los datos que está tratando de ingresar',
+                                'email' : email
+                            }
+                        )
+        
+            else:
+                return HttpResponseRedirect("/")
+           
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.EditarEmpleadoView %s method:post' % (self.template_name),
@@ -156,13 +198,18 @@ class EliminarEmpleadoView(TemplateView):
     
     def post(self, request):
         try:
-            if request.method == 'POST':
-                form = EliminarEmpleadoForm(request.POST)
-                if form.is_valid():
-                    pk = form.cleaned_data['pk']
-                    empleado = get_object_or_404(Empleado,pk = pk)
-                    empleado.delete()
-                    return HttpResponseRedirect('/empleados')
+            if 'email' in request.session:
+                
+                if request.method == 'POST':
+                    form = EliminarEmpleadoForm(request.POST)
+                    if form.is_valid():
+                        pk = form.cleaned_data['pk']
+                        empleado = get_object_or_404(Empleado,pk = pk)
+                        empleado.delete()
+                        return HttpResponseRedirect('/empleados')
+            else:
+                return HttpResponseRedirect('/')
+            
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.EliminarEmpleadoView %s method:post' % (self.template_name),
@@ -173,18 +220,28 @@ class ReporteEmpleadosPDF(View):
 
     def get(self, request):
         try:
-            empleados = Empleado.objects.order_by('-fecha_registro')
-            fecha = timezone.now()
-            correo = 'correo'
-            params = {
-                'email' : correo,
-                'fecha' : fecha,
-                'empleados': empleados,
-            }
-            return Render.render('render_pdf/rpt_empleados.html', params)
+            if 'email' in request.session:
+                empleados = Empleado.objects.order_by('-fecha_registro')
+                fecha = timezone.now()
+                correo = 'correo'
+                params = {
+                    'email' : correo,
+                    'fecha' : fecha,
+                    'empleados': empleados,
+                }
+                return Render.render('render_pdf/rpt_empleados.html', params)
+            else:
+                return HttpResponseRedirect("/")
+
         except Exception as e:
             excepcion.enviar_mensaje(
                 'crud.views.ReporteEmpleadoView %s method:ge' % (self.template_name),
                 '%s' % (str(e))
             )
 
+def salir(request):
+    if 'email' in request.session:
+        del request.session['email']
+        return HttpResponseRedirect("/")
+    else:
+        return HttpResponseRedirect("/empleados/")
